♻️ Move SpatialDataCurator from _spatial into __init__.py
